move interface to util
