+++ weldx/asdf/util.py
@@ -1,9 +1,9 @@
 """Utilities for asdf files."""
-from collections.abc import Mapping
 from distutils.version import LooseVersion
 from io import BytesIO
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Tuple, Type, Union
+from typing import Any, Callable, Dict, List, Tuple, Type, Union, MutableMapping, \
+    AbstractSet, Mapping, Hashable
 from warnings import warn
 
 import asdf
@@ -38,6 +38,7 @@ __all__ = [
     "view_tree",
     "notebook_fileprinter",
     "dataclass_serialization_class",
+    "_PROTECTED_KEYS"
 ]
 
 
@@ -551,3 +552,73 @@ def _get_instance_shape(
         if hasattr(converter, "shape_from_tagged"):
             return converter.shape_from_tagged(instance_dict)
     return None
+
+
+class _ProtectedViewDict(MutableMapping):
+    def __init__(self, protected_keys, data=None):
+        super(_ProtectedViewDict, self).__init__()
+        self._data = data
+        self.protected_keys = protected_keys
+
+    def __len__(self) -> int:
+        return len(self.keys())
+
+    def __getitem__(self, key):
+        if key in self.protected_keys:
+            self._warn_protected_keys()
+            raise KeyError
+        return self._data.get(key)
+
+    def __delitem__(self, key):
+        if key in self.protected_keys:
+            self._warn_protected_keys()
+            return
+        del self._data[key]
+
+    def __setitem__(self, key, value):
+        if key in self.protected_keys:
+            self._warn_protected_keys()
+            return
+        self._data[key] = value
+
+    def keys(self) -> AbstractSet:
+        return {k for k in self._data.keys() if k not in self.protected_keys}
+
+    def __iter__(self):
+        return (k for k in self._data.keys())
+
+    def __contains__(self, item):
+        return item in self.keys()
+
+    def update(self, mapping: Mapping[Hashable, Any], **kwargs: Any):
+        _mapping = dict(mapping, **kwargs)  # merge mapping and kwargs
+        if any(key in self.protected_keys for key in _mapping.keys()):
+            self._warn_protected_keys()
+            _mapping = {k: v for k, v in _mapping.items()
+                        if k not in self.protected_keys}
+
+        self._data.update(_mapping)
+
+    def popitem(self) -> Tuple[Hashable, Any]:
+        for k in self.keys():
+            if k in self.protected_keys:
+                continue
+
+            return k, self.pop(k)
+
+        raise KeyError
+
+    def clear(self):
+        _protected_data = {k: self._data.pop(k) for k in self.protected_keys}
+        self._data.clear()
+        self._data.update(_protected_data)  # re-add protected data.
+        assert len(self) == 0
+
+    def _warn_protected_keys(self):
+        import warnings
+
+        warnings.warn(
+            "You tried to manipulate an ASDF internal structure"
+            f" (currently protected: {self.protected_keys}",
+            stacklevel=3,
+        )
